+++ src/app/view/modal.rs
@@ -198,7 +198,6 @@ impl OpenCADStudio {
                         &self.default_save_format,
                         self.file_assoc_enabled,
                         self.write_dwg_native_constraints,
-                        self.show_constraint_values,
                         &self.ui_theme,
                         &self.theme_color_inputs,
                         self.language,